i2c: enable i2c-6 on non-acpi version
This enables a hack patch the lets the SCU change the pinmux for i2c-6. In order for the SCU driver
to be probed before the i2c driver, the i2c driver is built as a modules.
The patch is only applied to the kernel when DISTRO_FEATURES 'acpi' is not set.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>